requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
chromadb
pyspark
fastapi
uvicorn
requests
beautifulsoup4
findspark
pysqlite3-binary
setuptools
pandas
sentence-transformers
Pillow
pytesseract
